+++ core-tests/src/test/java/org/mule/runtime/core/transformer/simple/AddFlowVariableProcessorTestCase.java
@@ -16,7 +16,7 @@ import static org.junit.Assert.assertThat;
 import org.mule.functional.transformer.simple.AbstractAddVariablePropertyProcessorTestCase;
 import org.mule.runtime.api.metadata.DataType;
 import org.mule.runtime.core.api.MuleEvent;
-import org.mule.runtime.core.event.mutator.AddFlowVariableProcessor;
+import org.mule.runtime.core.processor.simple.AddFlowVariableProcessor;
 import org.mule.tck.size.SmallTest;
 
 import static org.hamcrest.collection.IsEmptyCollection.empty;
