+++ core-tests/src/test/java/org/mule/runtime/core/transformer/simple/RemovePropertyProcessorTestCase.java
@@ -15,7 +15,7 @@ import static org.mockito.Mockito.when;
 import org.mule.functional.transformer.simple.AbstractRemoveVariablePropertyProcessorTestCase;
 import org.mule.runtime.core.api.MuleEvent;
 import org.mule.runtime.core.api.MuleMessage;
-import org.mule.runtime.core.event.mutator.RemovePropertyProcessor;
+import org.mule.runtime.core.processor.simple.RemovePropertyProcessor;
 import org.mule.tck.size.SmallTest;
 
 import java.util.HashSet;
