+++ core/src/main/java/org/mule/runtime/core/exception/TemplateOnErrorHandler.java
@@ -109,8 +109,6 @@ public abstract class TemplateOnErrorHandler extends AbstractExceptionListener
         logger.warn(ex.getMessage());
       }
 
-      // event.setMessage(MuleMessage.builder(event.getMessage()).exceptionPayload(new
-      // DefaultExceptionPayload(exception)).build());
       throw exception;
     }
