+++ core/src/main/java/org/mule/runtime/core/processor/simple/RemoveFlowVariableProcessor.java
@@ -4,20 +4,27 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.event.mutator;
+package org.mule.runtime.core.processor.simple;
 
-import org.mule.runtime.api.metadata.DataType;
+import org.mule.runtime.core.PropertyScope;
 import org.mule.runtime.core.api.MuleEvent;
 
-public class AddFlowVariableProcessor extends AbstractAddVariablePropertyProcessor<Object> {
+import java.util.Set;
+
+public class RemoveFlowVariableProcessor extends AbstractRemoveVariablePropertyProcessor {
 
   @Override
-  protected MuleEvent addProperty(MuleEvent event, String propertyName, Object value, DataType dataType) {
-    return MuleEvent.builder(event).addFlowVariable(propertyName, value, dataType).build();
+  protected Set<String> getPropertyNames(MuleEvent event) {
+    return event.getFlowVariableNames();
   }
 
   @Override
   protected MuleEvent removeProperty(MuleEvent event, String propertyName) {
     return MuleEvent.builder(event).removeFlowVariable(propertyName).build();
+  }
+
+  @Override
+  protected String getScopeName() {
+    return PropertyScope.FLOW_VAR_NAME;
   }
 }
