+++ core/src/main/java/org/mule/runtime/core/processor/simple/SimpleMessageProcessor.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.event.mutator;
+package org.mule.runtime.core.processor.simple;
 
 import org.mule.runtime.core.AbstractAnnotatedObject;
 import org.mule.runtime.core.api.MuleContext;
@@ -15,12 +15,12 @@ import org.mule.runtime.core.api.lifecycle.Initialisable;
 import org.mule.runtime.core.api.processor.MessageProcessor;
 
 /**
- * Base clkass for {@link MessageProcessor}s that change the event or message. Implementations will return a new instance of the
+ * Base class for {@link MessageProcessor}s that change the event or message. Implementations will return a new instance of the
  * event with the updated data, since the {@link MuleEvent} and {@link MuleMessage} objects are immutable.
  *
  * @since 4.0
  */
-public abstract class AbstractEventMutatorProcessor extends AbstractAnnotatedObject
+public abstract class SimpleMessageProcessor extends AbstractAnnotatedObject
     implements MessageProcessor, MuleContextAware, Initialisable {
 
   protected MuleContext muleContext;
