+++ core/src/main/resources/META-INF/mule-module.properties
@@ -66,7 +66,6 @@ artifact.export.classPackages=org.mule.runtime.core,\
                               org.mule.runtime.core.el.mvel,\
                               org.mule.runtime.core.el.mvel.datatype,\
                               org.mule.runtime.core.enricher,\
-                              org.mule.runtime.core.event.mutator,\
                               org.mule.runtime.core.exception,\
                               org.mule.runtime.core.execution,\
                               org.mule.runtime.core.expression,\
@@ -90,6 +89,7 @@ artifact.export.classPackages=org.mule.runtime.core,\
                               org.mule.runtime.core.object,\
                               org.mule.runtime.core.processor,\
                               org.mule.runtime.core.processor.chain,\
+                              org.mule.runtime.core.processor.simple,\
                               org.mule.runtime.core.processor.strategy,\
                               org.mule.runtime.core.registry,\
                               org.mule.runtime.core.retry,\
