+++ modules/spring-config/src/main/java/org/mule/runtime/config/spring/dsl/model/CoreComponentBuildingDefinitionProvider.java
@@ -102,12 +102,6 @@ import org.mule.runtime.core.construct.Flow;
 import org.mule.runtime.core.context.notification.ListenerSubscriptionPair;
 import org.mule.runtime.core.context.notification.ServerNotificationManager;
 import org.mule.runtime.core.enricher.MessageEnricher;
-import org.mule.runtime.core.event.mutator.AbstractAddVariablePropertyProcessor;
-import org.mule.runtime.core.event.mutator.AddFlowVariableProcessor;
-import org.mule.runtime.core.event.mutator.AddPropertyProcessor;
-import org.mule.runtime.core.event.mutator.RemoveFlowVariableProcessor;
-import org.mule.runtime.core.event.mutator.RemovePropertyProcessor;
-import org.mule.runtime.core.event.mutator.SetPayloadMessageProcessor;
 import org.mule.runtime.core.exception.DefaultMessagingExceptionStrategy;
 import org.mule.runtime.core.exception.ErrorHandler;
 import org.mule.runtime.core.exception.OnErrorContinueHandler;
@@ -133,6 +127,12 @@ import org.mule.runtime.core.processor.AsyncDelegateMessageProcessor;
 import org.mule.runtime.core.processor.IdempotentRedeliveryPolicy;
 import org.mule.runtime.core.processor.ResponseMessageProcessorAdapter;
 import org.mule.runtime.core.processor.TransactionalMessageProcessor;
+import org.mule.runtime.core.processor.simple.AbstractAddVariablePropertyProcessor;
+import org.mule.runtime.core.processor.simple.AddFlowVariableProcessor;
+import org.mule.runtime.core.processor.simple.AddPropertyProcessor;
+import org.mule.runtime.core.processor.simple.RemoveFlowVariableProcessor;
+import org.mule.runtime.core.processor.simple.RemovePropertyProcessor;
+import org.mule.runtime.core.processor.simple.SetPayloadMessageProcessor;
 import org.mule.runtime.core.routing.AggregationStrategy;
 import org.mule.runtime.core.routing.ChoiceRouter;
 import org.mule.runtime.core.routing.FirstSuccessful;
