+++ modules/spring-config/src/main/java/org/mule/runtime/config/spring/handlers/MuleNamespaceHandler.java
@@ -98,11 +98,6 @@ import org.mule.runtime.core.context.notification.ListenerSubscriptionPair;
 import org.mule.runtime.core.el.ExpressionLanguageComponent;
 import org.mule.runtime.core.enricher.MessageEnricher;
 import org.mule.runtime.core.enricher.MessageEnricher.EnrichExpressionPair;
-import org.mule.runtime.core.event.mutator.AddFlowVariableProcessor;
-import org.mule.runtime.core.event.mutator.AddPropertyProcessor;
-import org.mule.runtime.core.event.mutator.RemoveFlowVariableProcessor;
-import org.mule.runtime.core.event.mutator.RemovePropertyProcessor;
-import org.mule.runtime.core.event.mutator.SetPayloadMessageProcessor;
 import org.mule.runtime.core.exception.DefaultMessagingExceptionStrategy;
 import org.mule.runtime.core.exception.ErrorHandler;
 import org.mule.runtime.core.exception.OnErrorContinueHandler;
@@ -126,6 +121,11 @@ import org.mule.runtime.core.object.PrototypeObjectFactory;
 import org.mule.runtime.core.object.SingletonObjectFactory;
 import org.mule.runtime.core.processor.IdempotentRedeliveryPolicy;
 import org.mule.runtime.core.processor.InvokerMessageProcessor;
+import org.mule.runtime.core.processor.simple.AddFlowVariableProcessor;
+import org.mule.runtime.core.processor.simple.AddPropertyProcessor;
+import org.mule.runtime.core.processor.simple.RemoveFlowVariableProcessor;
+import org.mule.runtime.core.processor.simple.RemovePropertyProcessor;
+import org.mule.runtime.core.processor.simple.SetPayloadMessageProcessor;
 import org.mule.runtime.core.processor.strategy.AsynchronousProcessingStrategy;
 import org.mule.runtime.core.processor.strategy.NonBlockingProcessingStrategy;
 import org.mule.runtime.core.retry.notifiers.ConnectNotifier;
