+++ transports/core/src/main/java/org/mule/compatibility/core/processor/simple/AddAttachmentProcessor.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.compatibility.core.event.mutator;
+package org.mule.compatibility.core.processor.simple;
 
 import static org.mule.runtime.core.util.IOUtils.toDataHandler;
 
@@ -14,8 +14,8 @@ import org.mule.runtime.core.api.MuleEvent;
 import org.mule.runtime.core.api.MuleException;
 import org.mule.runtime.core.api.MuleMessage;
 import org.mule.runtime.core.api.lifecycle.InitialisationException;
-import org.mule.runtime.core.event.mutator.AbstractEventMutatorProcessor;
 import org.mule.runtime.core.exception.MessagingException;
+import org.mule.runtime.core.processor.simple.SimpleMessageProcessor;
 import org.mule.runtime.core.util.AttributeEvaluator;
 
 import java.text.MessageFormat;
@@ -23,7 +23,7 @@ import java.text.MessageFormat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class AddAttachmentProcessor extends AbstractEventMutatorProcessor {
+public class AddAttachmentProcessor extends SimpleMessageProcessor {
 
   private static final Logger logger = LoggerFactory.getLogger(AddAttachmentProcessor.class);
