+++ transports/core/src/main/java/org/mule/compatibility/core/processor/simple/AddSessionVariableProcessor.java
@@ -4,11 +4,11 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.compatibility.core.event.mutator;
+package org.mule.compatibility.core.processor.simple;
 
 import org.mule.runtime.api.metadata.DataType;
 import org.mule.runtime.core.api.MuleEvent;
-import org.mule.runtime.core.event.mutator.AbstractAddVariablePropertyProcessor;
+import org.mule.runtime.core.processor.simple.AbstractAddVariablePropertyProcessor;
 
 import java.io.Serializable;
