+++ transports/core/src/main/java/org/mule/compatibility/core/processor/simple/RemoveAttachmentProcessor.java
@@ -4,22 +4,22 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.compatibility.core.event.mutator;
+package org.mule.compatibility.core.processor.simple;
 
 import org.mule.runtime.core.api.MuleEvent;
 import org.mule.runtime.core.api.MuleException;
 import org.mule.runtime.core.api.MuleMessage;
 import org.mule.runtime.core.api.MuleMessage.Builder;
 import org.mule.runtime.core.api.lifecycle.InitialisationException;
-import org.mule.runtime.core.event.mutator.AbstractEventMutatorProcessor;
 import org.mule.runtime.core.exception.MessagingException;
+import org.mule.runtime.core.processor.simple.SimpleMessageProcessor;
 import org.mule.runtime.core.util.AttributeEvaluator;
 import org.mule.runtime.core.util.WildcardAttributeEvaluator;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class RemoveAttachmentProcessor extends AbstractEventMutatorProcessor {
+public class RemoveAttachmentProcessor extends SimpleMessageProcessor {
 
   private static final Logger logger = LoggerFactory.getLogger(RemoveAttachmentProcessor.class);
