+++ transports/core/src/main/java/org/mule/compatibility/core/processor/simple/SetCorrelationIdTransformer.java
@@ -4,16 +4,16 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.compatibility.core.event.mutator;
+package org.mule.compatibility.core.processor.simple;
 
 import org.mule.runtime.core.DefaultMuleEvent;
 import org.mule.runtime.core.api.MuleEvent;
 import org.mule.runtime.core.api.MuleException;
 import org.mule.runtime.core.api.lifecycle.InitialisationException;
-import org.mule.runtime.core.event.mutator.AbstractEventMutatorProcessor;
+import org.mule.runtime.core.processor.simple.SimpleMessageProcessor;
 import org.mule.runtime.core.util.AttributeEvaluator;
 
-public class SetCorrelationIdTransformer extends AbstractEventMutatorProcessor {
+public class SetCorrelationIdTransformer extends SimpleMessageProcessor {
 
   private AttributeEvaluator correlationIdEvaluator;
