+++ transports/core/src/main/resources/META-INF/mule-module.properties
@@ -24,7 +24,7 @@ artifact.export.classPackages=org.mule.compatibility.core,\
                               org.mule.compatibility.core.endpoint.dynamic,\
                               org.mule.compatibility.core.endpoint.inbound,\
                               org.mule.compatibility.core.endpoint.outbound,\
-                              org.mule.compatibility.core.event.mutator,\
+                              org.mule.compatibility.core.processor.simple,\
                               org.mule.compatibility.core.exception,\
                               org.mule.compatibility.core.message,\
                               org.mule.compatibility.core.registry,\
