+++ transports/module-support/src/main/java/org/mule/compatibility/config/spring/handlers/MuleTransportsNamespaceHandler.java
@@ -21,16 +21,16 @@ import org.mule.compatibility.core.component.DefaultJavaWithBindingComponent;
 import org.mule.compatibility.core.component.PooledJavaWithBindingsComponent;
 import org.mule.compatibility.core.config.ConnectorConfiguration;
 import org.mule.compatibility.core.endpoint.EndpointURIEndpointBuilder;
-import org.mule.compatibility.core.event.mutator.AddAttachmentProcessor;
-import org.mule.compatibility.core.event.mutator.AddSessionVariableProcessor;
-import org.mule.compatibility.core.event.mutator.CopyAttachmentsProcessor;
-import org.mule.compatibility.core.event.mutator.RemoveAttachmentProcessor;
-import org.mule.compatibility.core.event.mutator.RemoveSessionVariableProcessor;
-import org.mule.compatibility.core.event.mutator.SetCorrelationIdTransformer;
+import org.mule.compatibility.core.processor.simple.AddAttachmentProcessor;
+import org.mule.compatibility.core.processor.simple.AddSessionVariableProcessor;
+import org.mule.compatibility.core.processor.simple.CopyAttachmentsProcessor;
+import org.mule.compatibility.core.processor.simple.RemoveAttachmentProcessor;
+import org.mule.compatibility.core.processor.simple.RemoveSessionVariableProcessor;
+import org.mule.compatibility.core.processor.simple.SetCorrelationIdTransformer;
 import org.mule.compatibility.core.routing.EndpointDlqUntilSuccessful;
 import org.mule.compatibility.core.routing.outbound.ExpressionRecipientList;
 import org.mule.compatibility.core.routing.requestreply.SimpleAsyncEndpointRequestReplyRequester;
-import org.mule.compatibility.core.transformer.simple.MutatorTransformerWrapper;
+import org.mule.compatibility.core.transformer.simple.MessageProcessorTransformerAdaptor;
 import org.mule.compatibility.module.cxf.builder.WebServiceMessageProcessorWithInboundEndpointBuilder;
 import org.mule.compatibility.module.cxf.component.WebServiceWrapperComponent;
 import org.mule.compatibility.module.cxf.config.JaxWsClientWithDecoupledEndpointFactoryBean;
@@ -69,7 +69,9 @@ public class MuleTransportsNamespaceHandler extends AbstractMuleNamespaceHandler
     registerBeanDefinitionParser("default-exception-strategy",
                                  new ExceptionStrategyDefinitionParser(DefaultMessagingExceptionStrategy.class));
 
-    registerBeanDefinitionParser("mutator-transformer", new MessageProcessorDefinitionParser(MutatorTransformerWrapper.class));
+    // TODO MULE-10457 Remove this element and perform the wrapping transparently
+    registerBeanDefinitionParser("mutator-transformer",
+                                 new MessageProcessorDefinitionParser(MessageProcessorTransformerAdaptor.class));
 
     registerMuleBeanDefinitionParser("set-session-variable",
                                      new MessageProcessorWithDataTypeDefinitionParser(AddSessionVariableProcessor.class))
